Disable sky clip behavior for skewed mid-textures
Ideally, this would alter the skew coordinates to keep the clip behavior and still allow skewing. I couldn't get that to work, so I've simply disabled it for skewed textures. I think this is a reasonable compromise because the clip behavior is just emulating the software renderer, and this is a feature that's only for the hardware renderer. (It's a lot better than these textures being *completely* broken, at least.)