Allow removing status screen by setting to "None"
Currently my mod has a dummy blank status screen that ends immediately, because it tries to never go to intermission, but there's some cases where it's forced (like the `changemap` command). This is a problem for the new multiplayer ready, since it's no longer a quick and seamless transition. Thought about maybe adding more control over the ready system, but my use case is a gross hack and I don't really need it for any other cutscenes. So let's just add the actual ability for a game to remove intermissions.